requirements.txt
python-decouple==3.6
requests==2.27.1
selenium==4.1.5
